Divide different tests into different classes
