+++ tests/Unit/BehaviorRemovingStrategyTest.php
@@ -6,23 +6,14 @@ namespace Yiisoft\Yii\Queue\Tests\Unit;
 
 use PHPUnit\Framework\Assert;
 use Yiisoft\Yii\Queue\FailureStrategy\BehaviorRemovingStrategy;
-use Yiisoft\Yii\Queue\FailureStrategy\ExponentialDelayStrategy;
 use Yiisoft\Yii\Queue\FailureStrategy\FailureStrategyInterface;
 use Yiisoft\Yii\Queue\FailureStrategy\PipelineInterface;
-use Yiisoft\Yii\Queue\FailureStrategy\SendAgainStrategy;
 use Yiisoft\Yii\Queue\Message\Message;
 use Yiisoft\Yii\Queue\Message\MessageInterface;
-use Yiisoft\Yii\Queue\Payload\PayloadInterface;
-use Yiisoft\Yii\Queue\PayloadFactory;
-use Yiisoft\Yii\Queue\Queue;
 use Yiisoft\Yii\Queue\Tests\TestCase;
 
 class BehaviorRemovingStrategyTest extends TestCase
 {
-    private const EXPONENTIAL_STRATEGY_DELAY_INITIAL = 1;
-    private const EXPONENTIAL_STRATEGY_DELAY_MAXIMUM = 5;
-    private const EXPONENTIAL_STRATEGY_EXPONENT = 2;
-
     public function strategyProvider(): array
     {
         return [
@@ -50,124 +41,6 @@ class BehaviorRemovingStrategyTest extends TestCase
                 new BehaviorRemovingStrategy('testKey', 'non-existing'),
                 ['testKey' => 'testValue', 'testKey2' => 'testValue2'],
                 ['testKey2' => 'testValue2'],
-            ],
-        ];
-    }
-
-    public function queueSendingStrategyProvider(): array
-    {
-        return [
-            [
-                SendAgainStrategy::class,
-                true,
-                [],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 1],
-            ],
-            [
-                SendAgainStrategy::class,
-                true,
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 1],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 2],
-            ],
-            [
-                SendAgainStrategy::class,
-                false,
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 2],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 2],
-            ],
-            [
-                SendAgainStrategy::class,
-                true,
-                [SendAgainStrategy::META_KEY_RESEND . '-' => -1],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 1],
-            ],
-            [
-                SendAgainStrategy::class,
-                true,
-                [SendAgainStrategy::META_KEY_RESEND . '-' => -100],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 1],
-            ],
-            [
-                SendAgainStrategy::class,
-                false,
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 5],
-                [SendAgainStrategy::META_KEY_RESEND . '-' => 5],
-            ],
-
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_INITIAL * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                    PayloadInterface::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_INITIAL * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                ],
-            ],
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 1,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                ],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 1 * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 2,
-                    PayloadInterface::META_KEY_DELAY => 1 * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                ],
-            ],
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 2,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                ],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 2 * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 2,
-                    PayloadInterface::META_KEY_DELAY => 2 * self::EXPONENTIAL_STRATEGY_EXPONENT,
-                ],
-            ],
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                ],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 2,
-                    PayloadInterface::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                ],
-            ],
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 4,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                ],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 2,
-                    PayloadInterface::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                ],
-            ],
-            [
-                ExponentialDelayStrategy::class,
-                true,
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => 100,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 1,
-                ],
-                [
-                    ExponentialDelayStrategy::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                    ExponentialDelayStrategy::META_KEY_ATTEMPTS => 2,
-                    PayloadInterface::META_KEY_DELAY => self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                ],
             ],
         ];
     }
@@ -226,62 +99,5 @@ class BehaviorRemovingStrategyTest extends TestCase
         self::assertFalse($result);
     }
 
-    /**
-     * @dataProvider queueSendingStrategyProvider
-     *
-     * @param string $strategyName
-     * @param bool $suites
-     * @param array $metaInitial
-     * @param array $metaResult
-     */
-    public function testQueueSendingStrategies(
-        string $strategyName,
-        bool $suites,
-        array $metaInitial,
-        array $metaResult
-    ): void {
-        $pipelineAssertion = static function (MessageInterface $message) use ($metaResult) {
-            Assert::assertEquals($metaResult, $message->getPayloadMeta());
 
-            return false;
-        };
-        $pipeline = $this->createMock(PipelineInterface::class);
-        $pipeline->expects($suites ? self::never() : self::once())
-            ->method('handle')
-            ->willReturnCallback($pipelineAssertion);
-
-        $queueAssertion = static function (PayloadInterface $payload) use ($metaResult) {
-            Assert::assertEquals($metaResult, $payload->getMeta());
-
-            return null;
-        };
-        $queue = $this->createMock(Queue::class);
-        $queue->expects($suites ? self::once() : self::never())
-            ->method('push')
-            ->willReturnCallback($queueAssertion);
-
-        $strategy = $this->getStrategy($strategyName, $queue);
-
-        $message = new Message('test', null, $metaInitial);
-        $result = $strategy->handle($message, $pipeline);
-
-        self::assertEquals($suites, $result);
-    }
-
-    private function getStrategy(string $strategyName, Queue $queue): FailureStrategyInterface
-    {
-        switch ($strategyName) {
-            case SendAgainStrategy::class:
-                return new SendAgainStrategy('', 2, $queue, new PayloadFactory());
-            case ExponentialDelayStrategy::class:
-                return new ExponentialDelayStrategy(
-                    2,
-                    self::EXPONENTIAL_STRATEGY_DELAY_INITIAL,
-                    self::EXPONENTIAL_STRATEGY_DELAY_MAXIMUM,
-                    self::EXPONENTIAL_STRATEGY_EXPONENT,
-                    new PayloadFactory(),
-                    $queue
-                );
-        }
-    }
 }
